moving repository interfaces to separate package
