+++ src/main/java/com/example/waterdelivery/repository/OrderRepository.java
@@ -1,6 +1,6 @@
-package com.example.demospringboot.repository;
+package com.example.waterdelivery.repository;
 
-import com.example.demospringboot.domain.Order;
+import com.example.waterdelivery.domain.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OrderRepository extends JpaRepository<Order, Long> {
